함수 공통화
