+++ src/playground/blocks/block_ai_learning.js
@@ -25,8 +25,105 @@ const DropDownDynamicGenerator = {
             return [[Lang.Blocks.no_target, 'null']];
         }
     },
+    valueMap: () => {
+        const valueMap = Object.values(Entry.aiLearning.getTrainResult()?.valueMap || []);
+        if (valueMap?.length) {
+            return valueMap.map((name) => [name, name]);
+        } else {
+            return [[Lang.Blocks.no_target, 'null']];
+        }
+    },
 };
+
+const createParamBlock = ({
+    name,
+    length,
+    createFunc,
+    skeleton = 'basic_string_field',
+    params = [],
+    type,
+}) => {
+    const result = {};
+    new Array(length).fill(0).forEach((_, idx) => {
+        const index = idx + 1;
+        const blockName = `${name}_${index}`;
+        const EmptyArray = new Array(index * 2).fill(0);
+        const lastAttr = {
+            key: 1,
+            value: 1,
+        };
+        const paramsKeyMap = EmptyArray.reduce((acc, cur, idx, array) => {
+            if (idx % 2 !== 0) {
+                return acc;
+            }
+            const index = idx / 2 + 1;
+            acc[`ATTR${index}`] = idx + 1;
+            lastAttr.key = index;
+            lastAttr.value = idx + 1;
+            return acc;
+        }, {});
+        if (params.length) {
+            params.forEach((p, i) => {
+                paramsKeyMap[`ATTR${lastAttr.key + i + 1}`] = lastAttr.value + 1;
+            });
+        }
+        result[blockName] = {
+            color: EntryStatic.colorSet.block.default.AI_LEARNING,
+            outerLine: EntryStatic.colorSet.block.darken.AI_LEARNING,
+            skeleton,
+            statements: [],
+            params: [
+                ...EmptyArray.map((_, i) => {
+                    if (i % 2 !== 0) {
+                        return {
+                            type: 'Block',
+                            accept: 'string',
+                            defaultType: 'number',
+                        };
+                    }
+                    return {
+                        type: 'TextDynamic',
+                        setValue: () => {
+                            const table = Entry.aiLearning?.getTableData?.();
+                            if (table) {
+                                const {
+                                    select = [],
+                                    fields = [],
+                                } = Entry.aiLearning?.getTableData?.();
+                                return (
+                                    fields[select?.[0]?.[i / 2]] || Lang.AiLearning.model_attr_str
+                                );
+                            }
+                            return Lang.AiLearning.model_attr_str;
+                        },
+                    };
+                }),
+                ...params,
+            ],
+            events: {},
+            def: {
+                type: blockName,
+            },
+            pyHelpDef: {
+                params: [],
+                type: blockName,
+            },
+            paramsKeyMap,
+            class: 'ai_learning',
+            isNotFor: [`${type}_attr_${index}`],
+            func: createFunc(paramsKeyMap),
+            syntax: {
+                js: [],
+                py: [],
+            },
+        };
+    });
+    return result;
+};
+
 module.exports = {
+    DropDownDynamicGenerator,
+    createParamBlock,
     getBlocks() {
         return {
             learning_title_image: {
